Avoid cast

+++ llvm/lib/Transforms/Instrumentation/HWAddressSanitizer.cpp
@@ -1901,7 +1901,7 @@ void HWAddressSanitizer::ShadowMapping::init(Triple &TargetTriple,
   } else if (ClEnableKhwasan || InstrumentWithCalls) {
     InGlobal = false;
     InTls = false;
-    Offset = optOr(ClMappingOffset, (unsigned long)0);
+    Offset = optOr(ClMappingOffset, 0UL);
     WithFrameRecord = false;
   } else if (ClMappingOffset.getNumOccurrences() > 0) {
     InGlobal = false;
